use size_t data type

+++ source/falaise/snemo/datamodels/calibrated_data.cc
@@ -129,10 +129,10 @@ namespace snemo {
       // Calibrated calorimeter hits:
       out_ << indent << datatools::i_tree_dumpable::tag;
       out_ << "Calibrated calorimeter hits: " << _calibrated_calorimeter_hits_.size() << std::endl;
-      for (int i = 0; i < (int) _calibrated_calorimeter_hits_.size(); i++) {
+      for (size_t i = 0; i < _calibrated_calorimeter_hits_.size(); i++) {
         const calibrated_calorimeter_hit & calo_calib_hit = _calibrated_calorimeter_hits_.at(i).get();
         out_ << indent << datatools::i_tree_dumpable::skip_tag;
-        if ( i + 1 == (int) _calibrated_calorimeter_hits_.size()) {
+        if ( i + 1 == _calibrated_calorimeter_hits_.size()) {
           out_ << datatools::i_tree_dumpable::last_tag;
         } else {
           out_ << datatools::i_tree_dumpable::tag;
@@ -147,10 +147,10 @@ namespace snemo {
       // Calibrated tracker hits:
       out_ << indent << datatools::i_tree_dumpable::inherit_tag(inherit_);
       out_ << "Calibrated tracker hits: " << _calibrated_tracker_hits_.size() << std::endl;
-      for (int i = 0; i < (int) _calibrated_tracker_hits_.size(); i++) {
+      for (size_t i = 0; i < _calibrated_tracker_hits_.size(); i++) {
         const calibrated_tracker_hit & tracker_calib_hit = _calibrated_tracker_hits_.at(i).get();
         out_ << indent << datatools::i_tree_dumpable::inherit_skip_tag(inherit_);
-        if ( i + 1 == (int) _calibrated_tracker_hits_.size()) {
+        if ( i + 1 == _calibrated_tracker_hits_.size()) {
           out_ << datatools::i_tree_dumpable::last_tag;
         } else {
           out_ << datatools::i_tree_dumpable::tag;
